# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SchoolBox Offline API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, teacher, student]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/deactivate:
    patch:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── STUDENTS ──────────────────────────────────────────────────────
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: isActive
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}/attendance-summary:
    get:
      operationId: getStudentAttendanceSummary
      tags: [students]
      summary: Get student attendance summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fromDate
          in: query
          schema:
            type: ["string", "null"]
        - name: toDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Student attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAttendanceSummary"

  # ── CLASSES ──────────────────────────────────────────────────────
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      parameters:
        - name: gradeLevel
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}/lessons:
    get:
      operationId: listLessons
      tags: [subjects]
      summary: List lessons for a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lessons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [subjects]
      summary: Create lesson in subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [subjects]
      summary: Get lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

  # ── ATTENDANCE ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        classId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - role
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        classId:
          type: ["integer", "null"]
      required:
        - username
        - fullName
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        classId:
          type: ["integer", "null"]
        isActive:
          type: boolean

    # ── STUDENT ──────────────────────────────────────────────────────
    Student:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        nationalId:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [male, female, null]
        parentPhone:
          type: ["string", "null"]
        parentName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        gradeLevel:
          type: ["integer", "null"]
        isActive:
          type: boolean
        enrollmentDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - isActive
        - enrollmentDate
        - createdAt
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        nationalId:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [male, female, null]
        parentPhone:
          type: ["string", "null"]
        parentName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
      required:
        - firstName
        - lastName
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        nationalId:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [male, female, null]
        parentPhone:
          type: ["string", "null"]
        parentName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    StudentAttendanceSummary:
      type: object
      properties:
        studentId:
          type: integer
        totalSessions:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        excused:
          type: integer
        attendanceRatePct:
          type: number
        consecutiveAbsences:
          type: integer
        absentDates:
          type: array
          items:
            type: string
      required:
        - studentId
        - totalSessions
        - present
        - absent
        - late
        - excused
        - attendanceRatePct
        - consecutiveAbsences
        - absentDates

    # ── CLASS ─────────────────────────────────────────────────────────
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        gradeLevel:
          type: integer
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        lessonCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - code
        - name
        - gradeLevel
        - isActive
        - lessonCount
        - createdAt
    CreateSubjectBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        gradeLevel:
          type: integer
        description:
          type: ["string", "null"]
      required:
        - code
        - name
        - gradeLevel
    Lesson:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        durationMinutes:
          type: ["integer", "null"]
        fileType:
          type: ["string", "null"]
          enum: [pdf, video, html, audio, null]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - subjectId
        - title
        - orderIndex
        - createdAt
        - updatedAt
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        durationMinutes:
          type: ["integer", "null"]
        fileType:
          type: ["string", "null"]
          enum: [pdf, video, html, audio, null]
      required:
        - title
        - orderIndex

    # ── ATTENDANCE ────────────────────────────────────────────────────
...